refactor(T3): Suppression code mort + extraction ExerciseFormScreen
Phase 1 - Suppression code obsolète:
- lib/models.dart (modèles dupliqués avec lib/models/*.dart)
- lib/screens/exercises_screen.dart (doublon non utilisé)

Phase 2 - Extraction formulaire exercice:
- Nouveau: lib/screens/exercise_form_screen.dart (304L)
- Refactorisé: lib/screens/exercises_list_screen.dart (773L → 480L)

Architecture:
- Séparation claire liste vs formulaire
- Responsabilités isolées
- Fichiers < 500 lignes chacun
- Aucune régression fonctionnelle (compilation OK)

Tests: flutter analyze passé, warnings dépréciation uniquement
