+++ docs/specs/v0.4/v0.4_scope.md
@@ -6,14 +6,17 @@ Release technique, avec outillage, refactoring, changement d'architecture, authe
 
 | ID  | Nom feature / tâche principale                            | Lot         | Priorité  | Dépendances  | Statut      |
 |------|----------------------------------------------------------|-------------|-----------|--------------|-------------|
-| T1  | Intégration SonarCloud                                    | Technique   | Haute     | -            | Fait        |
-| T2  | Cahier de recette (tests manuels)                         | Technique   | Moyenne   | T1           | Fait        |
+| T1  | Intégration SonarCloud                                    | Technique   | Haute     | -            | **Fait**    |
+| T2  | Cahier de recette (tests manuels)                         | Technique   | Moyenne   | T1           | **Fait**    |
 | T3  | Refactoring sans régression                               | Technique   | Haute     | T1, T2       | En cours    |
 | T4  | Authentification via NexTarget-server                     | Technique   | Moyenne   | T3           | A faire     |
 | T5  | Déport des appels Mistral API vers NexTarget-server       | Technique   | Moyenne   | T3           | À faire     |
+| D1  | Dashboard harmonisation et graphiques multi-courbes       | Dashboard   | Haute     | -            | **Fait**    |
+| D2  | Graphiques évolution par méthode de prise (1/2 mains)     | Dashboard   | Haute     | D1           | **Fait**    |
+| D3  | Statistique "Prise dominante" (remplacement catégories)   | Dashboard   | Moyenne   | D1           | **Fait**    |
 | F1  | Règles de sécurité FFTir (révision contenu & affichage)   | Feature     | Basse     | -            | À faire     |
 | F2  | Duplication d'un Exercice                                 | Feature     | Basse     | -            | À faire     |
-| F3  | Graphe d'une Session réalisée                             | Feature     | Basse     | -            | À faire     |
+| F3  | Graphe d'une Session réalisée                             | Feature     | Basse     | D1           | À faire     |
 
 ## Ajouts Techniques
 
@@ -31,11 +34,16 @@ Release technique, avec outillage, refactoring, changement d'architecture, authe
    - Objectif : permettre à un humain de tester l’app après refactoring (vérification absence de régression).
 
 3. Refactoring sans régression
-   - Vérifier que les modèles présents dans `lib/models.dart` ne sont pas déjà dans `lib/models/*.dart`. Si c'est le cas, supprimer les modèles dans `lib/models.dart` sans perte de fonctionnalité ou erreur de compilation.
-   - S’assurer que les couches UI, business et data sont séparées (aucun appel réseau direct dans les widgets).
-   - Supprimer composants non utilisés (analyse statique).
-   - Scinder les fichiers > 500 lignes en sous-modules cohérents.
-   - Exécuter les tests unitaires existants après refactor.
+   - **✅ FAIT** : Couches UI, business et data séparées (aucun appel réseau direct dans les widgets)
+   - **✅ FAIT** : Dashboard refactorisé avec architecture modulaire (DashboardService, composants réutilisables)
+   - **🔴 CRITIQUE** : **Modèles dupliqués** - `lib/models.dart` contient des classes obsolètes dupliquées avec `lib/models/*.dart` (ShootingSession, Series)
+   - **🟡 EN COURS** : Fichiers > 500 lignes à découper :
+     - `exercises_list_screen.dart` (773L) - Extraire formulaire d'édition
+     - `session_form.dart` (663L) - Modulariser en sous-composants  
+     - `planned_session_wizard.dart` (572L) - Séparer étapes du wizard
+     - `session_detail_screen.dart` (556L) - Extraire sections en composants
+   - **🟡 À FAIRE** : Analyse statique pour supprimer composants non utilisés
+   - **🟡 À FAIRE** : Tests unitaires pour écrans volumineux après refactor
 
 4. Authentification via NexTarget-server
    - Utiliser le backend `NexTarget-server` comme provider d’authentification centralisée.
@@ -53,7 +61,58 @@ Release technique, avec outillage, refactoring, changement d'architecture, authe
    - Le client ne stocke ni clé ni prompt.
    - Sécurité : appels HTTPS + token JWT utilisateur.
 
-## Features
+## Réalisations Techniques Récentes
+
+### Architecture Dashboard (Octobre 2025)
+- **DashboardService** : Service centralisé pour agrégation des données dashboard
+  - Méthodes génériques pour évolution par `HandMethod`
+  - Séparation claire business logic / UI
+  - Couverture tests complète
+
+- **EvolutionChart Multi-courbes** :
+  - Support dual-axis avec normalisation automatique
+  - Architecture réutilisable (1 ou 2 courbes)
+  - Tooltips contextuels avancés
+  - Thématisation cohérente (couleurs par méthode de prise)
+
+- **Modèles de Données** :
+  - `AdvancedStatsData` : Refactorisation complète pour métriques pertinentes
+  - `EvolutionData` : Support multi-courbes et dual-axis
+  - `HandSpecificData` : Nouvelles données pour filtrage par méthode de prise
+
+### Métriques de Qualité
+- **Tests** : 16/16 dashboard tests passent, couverture étendue
+- **Architecture** : Séparation couches respectée, composants réutilisables
+- **Performance** : Optimisations calculs SMA et agrégations données
+
+## Dashboard (Améliorations Complétées)
+
+### D1 : Dashboard harmonisation et graphiques multi-courbes ✅ **FAIT**
+- **Objectif** : Refactorisation du composant `EvolutionChart` pour supporter les graphiques multi-courbes avec dual-axis
+- **Réalisations** :
+  - Architecture générique pour 1 ou 2 courbes simultanées
+  - Support dual-axis (pts à gauche, cm à droite) avec normalisation/dénormalisation automatique
+  - Tooltips avancés avec informations contextuelles (date, index série)
+  - Tests complets (16/16 passent)
+
+### D2 : Graphiques évolution par méthode de prise ✅ **FAIT**
+- **Objectif** : Ajouter des graphiques spécifiques aux méthodes de prise (1 main vs 2 mains)
+- **Réalisations** :
+  - "Scores et Groupement - 2 mains" (courbes amber/blue)
+  - "Scores et Groupement - 1 main" (courbes orangeAccent/redAccent)
+  - Service générique `generateHandMethodEvolutionData(HandMethod)` 
+  - Filtrage intelligent par `HandMethod` avec méthode de compatibilité
+  - Architecture DRY avec réutilisation complète du composant `EvolutionChart`
+
+### D3 : Statistique "Prise dominante" ✅ **FAIT**
+- **Objectif** : Remplacer "Catégorie dominante" par une métrique plus pertinente
+- **Réalisations** :
+  - Modèle `AdvancedStatsData` refactorisé (dominantHandMethod/dominantHandMethodPercentage)
+  - Logique de calcul basée sur toutes les séries utilisateur avec pourcentages précis
+  - Interface mise à jour : "1 main (XX.X%)" ou "2 mains (XX.X%)"
+  - Icône appropriée (back_hand) et tests mis à jour
+
+## Features (À Développer)
 
 ### Règles de sécurité
 
@@ -69,7 +128,7 @@ Release technique, avec outillage, refactoring, changement d'architecture, authe
    - Bouton "dupliquer" présent dans la vue Edition, à côté du bouton Enregistrer
 
 3. Feature : Graphe Score et Groupement dans la vue d'un Session
-   - Utilisation du composant générique 'EvolutionChart'
+   - **Prérequis** : Utilisation du composant générique `EvolutionChart` (déjà disponible ✅)
    - Données :
       - 2 courbes : Points + Groupement de chaque série de la session
       - Pas de tendance
